Added follows Backend also tweaked some endpoints

+++ SouthChandlerCycling/SouthChandlerCycling/Services/ISignupService.cs
@@ -6,16 +6,13 @@ namespace SouthChandlerCycling.Services
     public interface ISignupService
     {
         Signup AddSignup(SignupRequestData signupRequest);
-        //List<Signup> GetSignupsForRider(RiderRequestData RequestData);
-        //List<Signup> GetSignupsForRide(RiderRequestData RequestData);
-        //void DeleteSignUp(SignupRequestData signupRequest);
         bool IsAuthorizedRider(SignupRequestData RequestData);
         bool IsAuthorizedRider(int RiderId, string Authorization);
         bool IsAuthorizedAdmin(SignupRequestData RequestData);
         bool IsAuthorizedAdmin(int AdminId, string Authorization);
         bool IsAuthorizedRiderOrAdmin(SignupRequestData RequestData);
         bool IsAuthorizedRiderOrAdmin(int RequestingId, int TargetId, string Authorization);
-        bool RiderExists(int RiderId);
+        bool RiderExists(long RiderId);
         bool RideExists(int RideId);
         bool SignUpExists(int RiderId, int RideId);
     }
